+++ src/Facebook/GraphNodes/GraphEdge.php
@@ -32,8 +32,15 @@ use Facebook\Exceptions\FacebookSDKException;
  *
  * @package Facebook
  */
-class GraphEdge extends Collection
+class GraphEdge implements \ArrayAccess, \Countable, \IteratorAggregate
 {
+    /**
+     * The items contained in the collection.
+     *
+     * @var array
+     */
+    protected $items = [];
+
     /**
      * @var FacebookRequest The original request that generated this data.
      */
@@ -66,11 +73,44 @@ class GraphEdge extends Collection
     public function __construct(FacebookRequest $request, array $data = [], array $metaData = [], $parentEdgeEndpoint = null, $subclassName = null)
     {
         $this->request = $request;
+        $this->items = $data;
         $this->metaData = $metaData;
         $this->parentEdgeEndpoint = $parentEdgeEndpoint;
         $this->subclassName = $subclassName;
+    }
 
-        parent::__construct($data);
+    /**
+     * Get all of the items in the collection.
+     *
+     * @return array
+     */
+    public function all()
+    {
+        return $this->items;
+    }
+
+    /**
+     * Get the collection of items as a plain array.
+     *
+     * @return array
+     */
+    public function asArray()
+    {
+        return array_map(function ($value) {
+            return $value instanceof GraphEdge ? $value->asArray() : $value;
+        }, $this->items);
+    }
+
+    /**
+     * Run a map over each of the items.
+     *
+     * @param \Closure $callback
+     *
+     * @return static
+     */
+    public function map(\Closure $callback)
+    {
+        return new static($this->request, array_map($callback, $this->items), $this->metaData, $this->parentEdgeEndpoint, $this->subclassName);
     }
 
     /**
@@ -256,5 +296,100 @@ class GraphEdge extends Collection
         }
 
         return null;
+    }
+
+    /**
+     * Get the collection of items as JSON.
+     *
+     * @param int $options
+     *
+     * @return string
+     */
+    public function asJson($options = 0)
+    {
+        return json_encode($this->asArray(), $options);
+    }
+
+    /**
+     * Count the number of items in the collection.
+     *
+     * @return int
+     */
+    public function count()
+    {
+        return count($this->items);
+    }
+
+    /**
+     * Get an iterator for the items.
+     *
+     * @return \ArrayIterator
+     */
+    public function getIterator()
+    {
+        return new \ArrayIterator($this->items);
+    }
+
+    /**
+     * Determine if an item exists at an offset.
+     *
+     * @param mixed $key
+     *
+     * @return bool
+     */
+    public function offsetExists($key)
+    {
+        return array_key_exists($key, $this->items);
+    }
+
+    /**
+     * Get an item at a given offset.
+     *
+     * @param mixed $key
+     *
+     * @return mixed
+     */
+    public function offsetGet($key)
+    {
+        return $this->items[$key];
+    }
+
+    /**
+     * Set the item at a given offset.
+     *
+     * @param mixed $key
+     * @param mixed $value
+     *
+     * @return void
+     */
+    public function offsetSet($key, $value)
+    {
+        if (is_null($key)) {
+            $this->items[] = $value;
+        } else {
+            $this->items[$key] = $value;
+        }
+    }
+
+    /**
+     * Unset the item at a given offset.
+     *
+     * @param string $key
+     *
+     * @return void
+     */
+    public function offsetUnset($key)
+    {
+        unset($this->items[$key]);
+    }
+
+    /**
+     * Convert the collection to its string representation.
+     *
+     * @return string
+     */
+    public function __toString()
+    {
+        return $this->asJson();
     }
 }
